Add MIDI Off receive and send devices.
Receiver sends all note off and pedal commands, while the receiver
receives this commands, passing it on to the track that it is on.
This ensures that any notes that are active on tracks that become
unarmed, correctly stop when the key is released. This allows for
holding a chord, changing patch, and then starting playing the new patch
without any interruption.